requirements.txt
pandas
unidecode
mysql-connector-python
gspread